+++ changelog.md
@@ -1,6 +1,8 @@
 # Uiua Changelog
 
 ## Logpoint 3 - 2023-09-30
+### Language
+- Remove the restriction that all functions in a non-scalar function array all have the same signature
 ### Interpreter
 - Allow passing `--no-format` to `uiua watch`